requirements.txt
Flask==3.1.0
markitdown==0.0.1a3
python-dotenv==1.0.1

